news_dashboard/backend/config.py
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

DB_CONFIG = {
    "dbname": os.getenv("DB_NAME"),
    "user": os.getenv("DB_USER"),
    "password": os.getenv("DB_PASSWORD"),
    "host": os.getenv("DB_HOST"),
    "port": os.getenv("DB_PORT"),
}

if not DB_CONFIG["dbname"]:
    print("⚠️ ERROR: Database config not loaded properly!")
else:
    print("✅ Database config loaded!")
