news_dashboard/backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
python-dotenv
transformers
nltk
pandas
matplotlib
